[Minimax-M3] Upcast FP8 index queries for Triton dot

+++ vllm/models/minimax_m3/common/ops/index_topk.py
@@ -122,7 +122,7 @@ def _index_block_score_kernel(
         block_shape=(BLOCK_SIZE_Q, head_dim),
         order=(1, 0),
     )
-    q = tl.load(q_ptrs, boundary_check=(0,), padding_option="zero")
+    q = tl.load(q_ptrs, boundary_check=(0,), padding_option="zero").to(tl.bfloat16)
     q_start = prefix_len + pid_q * BLOCK_SIZE_Q
 
     off_q = tl.arange(0, BLOCK_SIZE_Q) + pid_q * BLOCK_SIZE_Q + prefix_len
@@ -145,7 +145,7 @@ def _index_block_score_kernel(
             + page * stride_ik_blk
             + off_k[None, :] * stride_ik_pos
             + off_d[:, None] * stride_ik_d,
-        ).to(q.dtype)  # upcast: the index cache may be fp8 (e4m3)
+        ).to(tl.bfloat16)
         qk = tl.dot(q, k, out_dtype=tl.float32)
         # apply causal mask as needed
         if q_start < i + BLOCK_SIZE_K:
@@ -359,7 +359,7 @@ def _decode_index_score_kernel(
         + off_d[:, None] * stride_q_d,
         mask=q_mask[None, :],
         other=0.0,
-    )  # [D,HQ]
+    ).to(tl.bfloat16)  # [D,HQ]
     for blk in tl.range(chunk_start_block, chunk_end_block):
         page = tl.load(bt_row + blk).to(tl.int64)
         pos = blk * BLOCK_SIZE_K + off_k
@@ -372,11 +372,9 @@ def _decode_index_score_kernel(
             + page * stride_ik_blk
             + off_k[:, None] * stride_ik_pos
             + off_d * stride_ik_d,
-        ).to(q.dtype)  # [N,D] (upcast: the index cache may be fp8/e4m3)
-        # The explicit upcast matches what Triton's mixed-input dot lowers to
-        # anyway (no fp8 x bf16 MMA exists), but also compiles on Triton
-        # front-ends that reject fp8 dot operands outright. fp32 accumulation
-        # keeps the per-block max score exact for the fp8 indexer.
+        ).to(tl.bfloat16)  # [N,D]
+        # BF16 operands keep FP8 dot instructions out of the fallback path;
+        # FP32 accumulation preserves score accuracy.
         kq = tl.dot(k, q, out_dtype=tl.float32)  # [N,HQ]
         kq = tl.where(pos_mask & q_mask[None, :], kq, float("-inf"))
         score = tl.max(kq, axis=0)  # [HQ]
